Move output tests inline so they can use private functions
