+++ src/graphs/output/graphviz.rs
@@ -7,11 +7,11 @@ use graphviz_rust::{
 };
 
 #[derive(Debug)]
-pub struct GraphStringBody {
-    pub graph_string: String,
+struct GraphStringBody {
+    graph_string: String,
 }
 
-pub fn graph_to_graphviz_body(g: &impl Digraph, color: &str, with_label: bool) -> GraphStringBody {
+fn graph_to_graphviz_body(g: &impl Digraph, color: &str, with_label: bool) -> GraphStringBody {
     let all_node_list: Vec<String> = (0..g.num_vertices()).map(|i| format!("{i}")).collect();
     let all_node_string = all_node_list.join("\n");
     let all_node_edges_list: Vec<String> = (0..g.num_vertices())
@@ -38,7 +38,7 @@ pub fn graph_to_graphviz_body(g: &impl Digraph, color: &str, with_label: bool) -
     }
 }
 
-pub fn path_to_graphviz_body(g: &impl Digraph, path: &GraphPath) -> GraphStringBody {
+fn path_to_graphviz_body(g: &impl Digraph, path: &GraphPath) -> GraphStringBody {
     let path_nodes_list: Vec<String> = (0..(path.path.len() - 1))
         .map(|i| {
             format!(
@@ -87,4 +87,78 @@ pub fn output_graph_to_file_with_path(g: &impl Digraph, path: &GraphPath, output
         ),
     };
     graph_string_to_file(graph_string_body, output_path);
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::geography::datastructures::LatLng;
+    use crate::graphs::datastructures::{ALDigraph, Digraph, GraphPath, NodeID, NodeIndex};
+    use crate::graphs::output::graphviz::{graph_to_graphviz_body, path_to_graphviz_body};
+
+    fn get_test_graph() -> ALDigraph {
+        let mut g = ALDigraph::new(5);
+
+        for i in 0..5 {
+            g.mut_nodes_data().add_node_data_by_parts(
+                NodeID(i),
+                LatLng {
+                    latitude: 0.0,
+                    longitude: 0.0,
+                },
+            );
+        }
+
+        g.add_edge_by_id(NodeID(1), NodeID(0), 1.0);
+        g.add_edge_by_id(NodeID(0), NodeID(2), 1.0);
+        g.add_edge_by_id(NodeID(2), NodeID(1), 1.0);
+        g.add_edge_by_id(NodeID(0), NodeID(3), 1.0);
+        g.add_edge_by_id(NodeID(3), NodeID(4), 1.0);
+
+        g
+    }
+
+    fn get_test_graph_path() -> GraphPath {
+        GraphPath {
+            path: vec![
+                NodeIndex(1),
+                NodeIndex(0),
+                NodeIndex(2),
+                NodeIndex(1),
+                NodeIndex(0),
+                NodeIndex(3),
+                NodeIndex(4),
+            ],
+        }
+    }
+
+    #[test]
+    fn graph_to_graphviz_body_test() {
+        assert_eq!(
+            graph_to_graphviz_body(&get_test_graph(), "black", true).graph_string,
+            r#"0
+1
+2
+3
+4
+0 -> 0[color="black",headlabel="1"]
+0 -> 1[color="black",headlabel="1"]
+1 -> 0[color="black",headlabel="1"]
+2 -> 0[color="black",headlabel="1"]
+3 -> 0[color="black",headlabel="1"]
+"#
+        );
+    }
+
+    #[test]
+    fn path_to_graphviz_body_test() {
+        assert_eq!(
+            path_to_graphviz_body(&get_test_graph(), &get_test_graph_path()).graph_string,
+            r#"1 -> 0[headlabel="1", color="red"]
+0 -> 2[headlabel="1", color="red"]
+2 -> 1[headlabel="1", color="red"]
+1 -> 0[headlabel="1", color="red"]
+0 -> 3[headlabel="1", color="red"]
+3 -> 4[headlabel="1", color="red"]"#
+        );
+    }
 }
