+++ tests/graphs/mod.rs
@@ -1,4 +1,3 @@
 mod connected_components_tests;
-mod output;
 mod shortest_path_tests;
 mod travelling_salesman_tests;
